AG-13216 Join the pickFocus implementation for linear/radial gauges
